build fix (2nd try)
svn:sync-xref-src-repo-rev: r59226

+++ src/VBox/Runtime/common/string/strformatrt.cpp
@@ -1,4 +1,4 @@
-/* $Id: strformatrt.cpp 27650 2010-03-23 23:07:08Z knut.osmundsen@oracle.com $ */
+/* $Id: strformatrt.cpp 27651 2010-03-23 23:10:53Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - IPRT String Formatter Extensions.
  */
@@ -1052,12 +1052,12 @@ size_t rtstrFormatRt(PFNRTSTROUTPUT pfnOutput, void *pvArgOutput, const char **p
                         static const char   s_szAttrEscape[] = "\n\""; /* more? */
                         static const char   s_szElemEscape[] = "<>&'\"";
                         const char *pszEscape = fAttr ?             s_szAttrEscape      :             s_szElemEscape;
-                        int         cchEscape = fAttr ? RT_ELEMENTS(s_szAttrEscape) - 1 : RT_ELEMENTS(s_szElemEscape) - 1;
+                        size_t      cchEscape = fAttr ? RT_ELEMENTS(s_szAttrEscape) - 1 : RT_ELEMENTS(s_szElemEscape) - 1;
                         size_t      cchOutput = 0;
                         const char *pszStr    = va_arg(*pArgs, char *);
-                        int         cchStr;
-                        int         offCur;
-                        int         offLast;
+                        ssize_t     cchStr;
+                        ssize_t     offCur;
+                        ssize_t     offLast;
 
                         if (!VALID_PTR(pszStr))
                             pszStr = "<NULL>";
